+++ com.unity.render-pipelines.high-definition/Runtime/Material/Decal/DecalProjector.cs
@@ -31,6 +31,10 @@ namespace UnityEngine.Rendering.HighDefinition
             }
         }
 
+#if UNITY_EDITOR
+        private int m_Layer;
+#endif
+
         [SerializeField]
         private float m_DrawDistance = 1000.0f;
         /// <summary>
@@ -224,6 +228,9 @@ namespace UnityEngine.Rendering.HighDefinition
             }
 
             Matrix4x4 sizeOffset = Matrix4x4.Translate(decalOffset) * Matrix4x4.Scale(decalSize);
+#if UNITY_EDITOR
+            m_Layer = gameObject.layer;
+#endif
             m_Handle = DecalSystem.instance.AddDecal(position, rotation, Vector3.one, sizeOffset, m_DrawDistance, m_FadeScale, uvScaleBias, m_AffectsTransparency, m_Material, gameObject.layer, m_FadeFactor);
         }
 
@@ -278,6 +285,18 @@ namespace UnityEngine.Rendering.HighDefinition
                 }
             }
         }
+
+#if UNITY_EDITOR
+        void Update() // only run in editor
+        {
+            if(m_Layer != gameObject.layer)
+            {
+                Matrix4x4 sizeOffset = Matrix4x4.Translate(decalOffset) * Matrix4x4.Scale(decalSize);
+                m_Layer = gameObject.layer;
+                DecalSystem.instance.UpdateCachedData(position, rotation, sizeOffset, m_DrawDistance, m_FadeScale, uvScaleBias, m_AffectsTransparency, m_Handle, gameObject.layer, m_FadeFactor);
+            }
+        }
+#endif
 
         void LateUpdate()
         {